# specs/001-bilibili-monitor/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bili Stats Monitor API
  version: v1
  description: |
    B站数据监控工具 API（MVP）。

    统一约定：
    - 仅使用 GET / POST。所有变更/删除等操作通过 POST + body.action 指定，不使用 PUT/PATCH/DELETE。
    - 统一响应包：{ code, message, data }
    - 时间：ISO 8601，时区 Asia/Shanghai（北京时间）
    - 分页：page（默认1），page_size（默认20，最大100）
    - 选择语义：
      - 批量接口 selection: { type: "ids"|"all", ids?: string[], filters?: object }
      - type=all 时作用于“当前筛选条件下的全部结果（跨页）”
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/accounts:
    get:
      tags: [Accounts]
      summary: 账号列表
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAccountList'
  /api/v1/accounts/default:
    get:
      tags: [Accounts]
      summary: 获取全局默认账号
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  message: { type: string }
                  data: { type: object, properties: { id: { type: string, nullable: true } } }
    post:
      tags: [Accounts]
      summary: 设置全局默认账号
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /api/v1/accounts/cookie:
    post:
      tags: [Accounts]
      summary: 绑定账号（Cookie）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cookie]
              properties:
                cookie: { type: string, description: '需包含 SESSDATA' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAccount'
  /api/v1/accounts/qrcode:
    post:
      tags: [Accounts]
      summary: 获取扫码登录二维码
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [create]
                  default: create
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      session_id: { type: string }
                      qr_url: { type: string }
                      expire_at: { type: string, format: date-time }
                      poll_interval_sec: { type: integer, default: 2 }
  /api/v1/accounts/qrcode/status:
    get:
      tags: [Accounts]
      summary: 轮询扫码登录状态
      parameters:
        - in: query
          name: session_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: [pending, scanned, confirmed, expired]
                      account:
                        $ref: '#/components/schemas/Account'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    page_size:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: 'ok' }
        data: { }
...
    ApiResponseAccount:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Account' }
    ApiResponseAccountList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/Account' }
                page: { type: integer }
                page_size: { type: integer }
                total: { type: integer }
...
    Account:
      type: object
      properties:
        id: { type: string }
        uid: { type: string }
        nickname: { type: string }
        bind_method: { type: string, enum: [cookie, qrcode] }
        status: { type: string, enum: [valid, expired] }
        last_failures: { type: integer }
        bound_at: { type: string, format: date-time }
...